tasks/object-detection/sam3-approach/pyproject.toml
[project]
name = "sam3-approach"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "accelerate>=1.13.0",
    "numpy>=2.4.3",
    "pillow>=12.1.1",
    "torch>=2.10.0",
    "torchvision>=0.25.0",
    "transformers>=4.49",
]
